Rename `exceptions.py` to `error_wrappers.py`
