+++ pydantic/__init__.py
@@ -1,6 +1,7 @@
 # flake8: noqa
+from . import errors
 from .env_settings import BaseSettings
-from .exceptions import *
+from .error_wrappers import ValidationError
 from .fields import Required
 from .main import BaseConfig, BaseModel, create_model, validator
 from .parse import Protocol
