+++ tests/test_validators.py
@@ -2,7 +2,7 @@ from typing import List
 
 import pytest
 
-from pydantic import BaseModel, ConfigError, ValidationError, validator
+from pydantic import BaseModel, ValidationError, errors, validator
 
 
 def test_simple():
@@ -191,7 +191,7 @@ def test_classmethod():
 
 
 def test_duplicates():
-    with pytest.raises(ConfigError) as exc_info:
+    with pytest.raises(errors.ConfigError) as exc_info:
         class Model(BaseModel):
             a: str
             b: str
@@ -208,7 +208,7 @@ def test_duplicates():
 
 
 def test_use_bare():
-    with pytest.raises(ConfigError) as exc_info:
+    with pytest.raises(errors.ConfigError) as exc_info:
         class Model(BaseModel):
             a: str
 
@@ -219,7 +219,7 @@ def test_use_bare():
 
 
 def test_use_no_fields():
-    with pytest.raises(ConfigError) as exc_info:
+    with pytest.raises(errors.ConfigError) as exc_info:
         class Model(BaseModel):
             a: str
 
@@ -320,7 +320,7 @@ def test_wildcard_validator_error():
 
 
 def test_invalid_field():
-    with pytest.raises(ConfigError) as exc_info:
+    with pytest.raises(errors.ConfigError) as exc_info:
         class Model(BaseModel):
             a: str
